+++ app/lib/auto-update/controller.js
@@ -9,29 +9,45 @@ module.exports = function(env, clientConfig, tray, about) {
   autoUpdater.logger = logger;
 
   autoUpdater.on('checking-for-update', () => {
-    logger.info('Autoupdater: Checking for update.');
+    logger.info('Checking for update', {category: 'autoupdate'});
     about.update('checking-for-update');
   });
 
   autoUpdater.on('update-available', (event, info) => {
-    logger.info('Autoupdater: Update available.', {info});
+    logger.info('update available', {
+      category: 'autoupdate',
+      data: info
+    });
+
     about.update('update-available');
   });
 
   autoUpdater.on('update-not-available', (event, info) => {
-    logger.info('Autoupdater: Update not available.', {info});
+    logger.info('Update not available', {
+      category: 'autoupdate',
+      data: info
+    });
+
     about.update('update-not-available');
     checkRunning = false;
   });
 
   autoUpdater.on('error', (event, err) => {
-    logger.error('Autoupdater: error', {err});
+    logger.error('error occured during autoupdate', {
+      category: 'autoupdate',
+      error: err
+    });
+
     about.update('error');
     checkRunning = false;
   });
 
   autoUpdater.on('update-downloaded', (event, info) => {
-    logger.info('Autoupdater: Update downloaded', {info});
+    logger.info('update has been downloaded', {
+      category: 'autoupdate',
+      data: info
+    });
+
     about.update('update-downloaded');
     tray.toggleState('update', true);
     clientConfig.set('updateAvailable', true);
@@ -39,7 +55,7 @@ module.exports = function(env, clientConfig, tray, about) {
   });
 
   function shouldCheckForUpdates() {
-    if(!process.defaultApp && env.name === 'production') { 
+    if(!process.defaultApp && env.name === 'production') {
       if(env.update && env.update.enable !== undefined) {
         return env.update.enable;
       } else {
@@ -55,7 +71,7 @@ module.exports = function(env, clientConfig, tray, about) {
       checkRunning = true;
       autoUpdater.checkForUpdates();
     } else {
-      logger.info('Autoupdate: skip check for update');
+      logger.info('skip check for update', {category: 'autoupdate'});
       about.update('update-not-available');
     }
   }
@@ -66,7 +82,7 @@ module.exports = function(env, clientConfig, tray, about) {
 
   function setUpdateCheckInterval() {
     if(shouldCheckForUpdates() === false) return;
-    
+
     var intervalD = 7;
     if(env.update && env.updateCheckInterval) {
       intervalD = env.update.updateCheckInterval;
@@ -74,7 +90,12 @@ module.exports = function(env, clientConfig, tray, about) {
 
     var intervalMs = intervalD * 24 * 60 * 60 * 1000;
 
-    logger.info('Autoupdate: setting update check interval', {intervalMs, intervalD});
+    logger.info('setting update check interval', {
+      category: 'autoupdate',
+      minutes: intervalMs,
+      days: intervalD
+    });
+
     setInterval(checkForUpdate, intervalMs);
   }